Bug fix migration

+++ src/Services/DynamicDBHelper.php
@@ -29,10 +29,11 @@ class DynamicDBHelper
             // ALTER TABLE
             $as_value = "json_unquote(json_extract(`value`,'$.$column_name'))";
 
-            DB::statement("ALTER TABLE $db_table_name ADD $db_column_name nvarchar(768) GENERATED ALWAYS AS ($as_value) VIRTUAL;");
-            DB::statement("ALTER TABLE $db_table_name ADD index $index_name($db_column_name)");
-
-            DB::commit();
+            if(\Schema::hasTable($db_table_name)){
+                DB::statement("ALTER TABLE $db_table_name ADD $db_column_name nvarchar(768) GENERATED ALWAYS AS ($as_value) VIRTUAL;");
+                DB::statement("ALTER TABLE $db_table_name ADD index $index_name($db_column_name)");    
+                DB::commit();
+            }
         } catch (Exception $exception) {
             DB::rollback();
             throw $exception;
@@ -43,13 +44,24 @@ class DynamicDBHelper
     {
         DB::beginTransaction();
         try {
-            // ALTER TABLE
-            DB::statement("ALTER TABLE $db_table_name DROP INDEX $index_name;");
-            DB::statement("ALTER TABLE $db_table_name DROP COLUMN $db_column_name;");
-            DB::commit();
+            if(\Schema::hasTable($db_table_name)){
+                // ALTER TABLE
+                DB::statement("ALTER TABLE $db_table_name DROP INDEX $index_name;");
+                DB::statement("ALTER TABLE $db_table_name DROP COLUMN $db_column_name;");
+                DB::commit();
+            }
         } catch (Exception $exception) {
             DB::rollback();
             throw $exception;
         }
+    }
+
+    public static function getDBIndex($db_table_name, $db_column_name, $unique = true){
+        if(!\Schema::hasTable($db_table_name)){
+            return collect([]);
+        }
+
+        $unique_key = boolval($unique) ? 0 : 1;
+        return \DB::select("SHOW INDEX FROM $db_table_name WHERE non_unique = $unique_key AND column_name = '$db_column_name'");
     }
 }
